[FIX] account_credit_control: Avoid error when opening or creating a mail template
When a user with the group_account_credit_control_manager group tries to select a mail template, an access error is raised because model_id has permissions restricted to Administration/Access Rights. Instead of using a many2one field, the model field is used, as it is related and stored.